These modifications get the code to compile.  This is the same version as source-willcompile.zip distributed on devel.hoverrace.com.  The vast majority of the changes here were changing the order in which Visual Studio compiled the pieces of the project.

+++ NetTarget/MazeCompiler/LevelBuilderVisiblesZones.cpp
@@ -31,7 +31,7 @@
 // Usefull constants
 #define MR_MAX_VISIBLE_ZONES   500
 
-static const double Pi = fabs( atan2( 0, -1 ));
+static const double Pi = fabs(atan2((double) 0, (double) -1 ));
 
 // Usefull local structures
 class MR_2DFloatPos
